requirements.txt
-e .
pytest
aws-cdk.aws_apigateway
aws-cdk.aws_lambda
aws-cdk.aws_certificatemanager

